refactor(test): move file into source from test
